# multiproof: improve performance of prover and verifier