--- a/common/transcript.go
+++ b/common/transcript.go
@@ -1,6 +1,7 @@
 package common
 
 import (
+	"bytes"
 	"crypto/sha256"
 	"hash"
 
@@ -12,6 +13,7 @@ import (
 // See: Fiat-Shamir
 type Transcript struct {
 	state hash.Hash
+	buff  *bytes.Buffer
 }
 
 func NewTranscript(label string) *Transcript {
@@ -20,21 +22,22 @@ func NewTranscript(label string) *Transcript {
 
 	transcript := &Transcript{
 		state: digest,
+		buff:  bytes.NewBuffer(make([]byte, 1<<20)),
 	}
 
 	return transcript
 }
 
-func (t *Transcript) AppendMessage(message []byte, label string) {
-	t.state.Write([]byte(label))
-	t.state.Write(message)
+func (t *Transcript) AppendMessage(message []byte, label []byte) {
+	t.buff.Write(label)
+	t.buff.Write(message)
 }
 
 // Appends a Scalar to the transcript
 //
 // Converts the scalar to 32 bytes, then appends it to
 // the state
-func (t *Transcript) AppendScalar(scalar *fr.Element, label string) {
+func (t *Transcript) AppendScalar(scalar *fr.Element, label []byte) {
 	tmpBytes := scalar.BytesLE()
 	t.AppendMessage(tmpBytes[:], label)
 
@@ -44,14 +47,14 @@ func (t *Transcript) AppendScalar(scalar *fr.Element, label string) {
 //
 // Compresses the Point into a 32 byte slice, then appends it to
 // the state
-func (t *Transcript) AppendPoint(point *banderwagon.Element, label string) {
+func (t *Transcript) AppendPoint(point *banderwagon.Element, label []byte) {
 	tmp_bytes := point.Bytes()
 	t.AppendMessage(tmp_bytes[:], label)
 
 }
 
-func (t *Transcript) DomainSep(label string) {
-	t.state.Write([]byte(label))
+func (t *Transcript) DomainSep(label []byte) {
+	t.buff.Write(label)
 }
 
 // Computes a challenge based off of the state of the transcript
@@ -60,9 +63,11 @@ func (t *Transcript) DomainSep(label string) {
 // scalar field
 //
 // Note that calling the transcript twice, will yield two different challenges
-func (t *Transcript) ChallengeScalar(label string) fr.Element {
+func (t *Transcript) ChallengeScalar(label []byte) fr.Element {
 	t.DomainSep(label)
 
+	t.state.Write(t.buff.Bytes())
+	t.buff.Reset()
 	// Reverse the endian so we are using little-endian
 	// SetBytes interprets the bytes in Big Endian
 	bytes := t.state.Sum(nil)