… common/transcript_test.go